feat: Add CSS to index.html, start work on lazy loading
A new method was added to the gulpfile. It will now take the css
for the index.html page and inject it into a style tag. This will
make it so the CSS will not render-block the page when it loads.
Additionaly, work was started on implementing lazy loading. All
image sources now use the extremely small version (generated by
the gulpfile) until I finish implementing lazy loading.