feat: Update form and input styling
Added a focus styling to buttons and inputs to look nicer and give
better visual feedback as to what is currently focused. Additionally
updated CSS of review submit button because I realized it was going
off the side of the review box when I didn't want it to do so.